feat: implementando rotas de requests

+++ api/models/budget.go
@@ -1,9 +1,9 @@
 package models
 
 type Budget struct {
-	ID     int   `json:"id"`
-	Year   int   `json:"year"`
-	Month  int   `json:"month"`
-	AreaID int   `json:"area_id"`
-	Amount int64 `json:"amount"`
+	ID     int `json:"id"`
+	Year   int `json:"year"`
+	Month  int `json:"month"`
+	AreaID int `json:"area_id"`
+	Amount int `json:"amount"`
 }
