+++ api/models/transaction.go
@@ -3,7 +3,7 @@ package models
 type Transaction struct {
 	ID         int    `json:"id"`
 	Date       string `json:"date"`
-	Amount     int64  `json:"amount"`
+	Amount     int    `json:"amount"`
 	CategoryID int    `json:"category_id"`
 	AreaID     int    `json:"area_id"`
 	Type       string `json:"type"` // 'income' or 'expense'
